requirements.txt
replit
discord.py
flask
python-dotenv==0.10.3